requirements.txt
# Django
asgiref==3.5.2
Django==3.2.14
pytz==2022.1
sqlparse==0.4.2
psycopg2~=2.9.3
django-environ~=0.9.0
phonenumbers==8.12.53
Pillow==9.2.0
django-debug-toolbar==3.5.0
celery~=5.2.7
Redis
django-redis~=5.2.0
django-celery-beat


# DRF
djangorestframework==3.13.1
dj-rest-auth==2.2.5
django-rest-authtoken==2.1.4
djangorestframework-simplejwt==5.2.0
drf-spectacular
drf-psq==1.1.0
django-extra-fields==3.0.2
django-filter==22.1


# Init scripts
Faker==14.1.0

